Reorganise code in to lib folder; seperate dependabot code from tinglesoftware code using unique module names
